+++ src/pages/Explore/ui/ProductDetail/ui/RelatedProducts/index.tsx
